refactor: extract TTL store and HTML builders from router.ts
Move createTtlStore into ttl-store.ts and HTML rendering helpers
(escapeHtml, buildLoginHtml, buildSignupHtml) into html-builders.ts
to reduce mixing of concerns in router.ts.
